Move Transaction & nodeInfo defs to dedicated files
